agrego cosas de admin

+++ src/public/js/adminhome.js
@@ -1,4 +1,5 @@
-document.addEventListener('DOMContentLoaded', function() {
+
+  document.addEventListener('DOMContentLoaded', function() {
     const deleteButtons = document.querySelectorAll('.btnDeleteProduct');
   
     deleteButtons.forEach(button => {
@@ -6,25 +7,27 @@ document.addEventListener('DOMContentLoaded', function() {
         const pid = button.getAttribute('data-product-id');
   
         if (pid) {
-            const confirmMessage = "Seguro deseas borrar el producto?";
-            const userConfirmed = window.confirm(confirmMessage)
-
-          fetch(`/api/products/deleteProduct/${pid}`, {
-            method: 'DELETE',
-            headers: {
-              'Content-Type':'application/json'
-            },
-            body: JSON.stringify({ pid: pid })
-          })
-          .then(response => response.json())
-          .then(data => {
-            if(data.status === "success"){
-              alert('Producto eliminado!')
-            }
-          })
-          .catch(error => {
-            console.error('Error:', error);
-          });
+          const confirmMessage = "Seguro deseas borrar el producto?";
+          const userConfirmed = window.confirm(confirmMessage);
+  
+          if (userConfirmed) {
+            fetch(`/api/products/deleteProduct/${pid}`, {
+              method: 'DELETE',
+              headers: {
+                'Content-Type': 'application/json'
+              },
+              body: JSON.stringify({ pid: pid })
+            })
+              .then(response => response.json())
+              .then(data => {
+                if (data.status === "success") {
+                  alert('Producto eliminado!')
+                }
+              })
+              .catch(error => {
+                console.error('Error:', error);
+              });
+          }
         }
       });
     });
